Client coordinates conversion
Worms are showing ok now.
There's a problem when there are worms too low on the stage, only the top of their head can be seen.
It's probably due to a mismatch between client and server coordinates (I think server sends the center coordinate while the client takes as it were a corner coordinate [not sure which corner SDL uses])